+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2024-07-17
+
 ### Added
 
 - ReclaimPolicy added in the Bucket CR to manage the data clean up (retain or delete).
@@ -121,7 +123,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement creation of S3 buckets on CAPA management clusters.
 
-[Unreleased]: https://github.com/giantswarm/object-storage-operator/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/giantswarm/object-storage-operator/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/giantswarm/object-storage-operator/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/giantswarm/object-storage-operator/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/giantswarm/object-storage-operator/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/giantswarm/object-storage-operator/compare/v0.5.5...v0.6.0